Beginnings of the Matrix model
The Tests for the Matrix model are defined in the Matrix.spec.ts file
The Matrix Model, its interface, and constructor interface are defined
in the Models/Matrix/index.ts file.

The matrix model is designed to be a 1D array from the start rather
than an array of arrays. The interface currently exposes a numRows and
numCols property and provides a get method to retrieve the values
from a specified row and column.

TDD is being used and currently all tests are passing with 100% coverage